backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
anthropic==0.28.0
pandas==2.2.2
openpyxl==3.1.4
pdfplumber==0.11.0
yfinance==0.2.40
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.0
simpleeval==0.9.13
numpy==1.26.4
